Fix up tracing propagator
